1:为 tcp service 添加会话的接口添加强制使用当前线程内的event loop 2:修改http服务接口

+++ src/examples/BroadCastClient.cpp
@@ -70,7 +70,7 @@ int main(int argc, char** argv)
                 }
 
                 /*  可以放入消息队列，然后唤醒它主线程的eventloop，然后主线程通过消息队列去获取*/
-                ds->setDataCallback([&total_recv, &packet_num](DataSocket::PTR ds, const char* buffer, int len){
+                ds->setDataCallback([&total_recv, &packet_num](DataSocket::PTR ds, const char* buffer, size_t len){
                     const char* parse_str = buffer;
                     int total_proc_len = 0;
                     int left_len = len;
